service-ids-check: Add check uid/gid of services
This new test allows to check whether service
run with expected uid and gid. This check doesn't
check that the service runs because there is an
other check for that.

It currently only checks weston service to run as
display/display but it can be easily extended for
checking other services.

Bug-AGL: SPEC-546

Signed-off-by: jobol <jose.bollo@iot.bzh>